Updated AddShipmentManually
On the TransportationAdmin form the AddShipmentManually button now works and creates a new shipment adding it to the correct warehouse. Also the AddShipmentFromFile form is working when you click JSON. It imports the shipments from the JSON file and adds them to the specified warehouses list of shipments. 
- one thing is though it would not work to import shipments with warehouse ids that do not exist and have a name
- it wouldnt work for any JSON files with a name field for the warehouse - something we need to think about
STILL NEEDED - ImportShipmentsXML functionality, Exit button to close the program, automatic window sizing is needed as well for when the new forms open from the button clicks on TransportationAdmin